pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rainwords"
version = "0.1.0"
description = "RainWords - AI-assisted poetry console"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }

dependencies = [
  "sentence-transformers",
  "faiss-cpu",
  "nltk",
  "numpy",
  "fastapi",
  "uvicorn[standard]",
  "pydantic",
  "pypdf",
  "spacy>=3.7",
  "fr_core_news_md @ https://github.com/explosion/spacy-models/releases/download/fr_core_news_md-3.7.0/fr_core_news_md-3.7.0-py3-none-any.whl",
]

[project.scripts]
rainwords = "rainwords.__main__:main"
"rainwords.corpus_builder" = "rainwords.corpus_builder:main"
"rainwords.convert_pdf"   = "rainwords.convert_pdf:main"


[tool.setuptools]
packages = ["rainwords"]
